Move polynomial into secret_sharing module
