+++ src/cryptographic_primitives/secret_sharing/mod.rs
@@ -6,3 +6,6 @@
 */
 
 pub mod feldman_vss;
+mod polynomial;
+
+pub use polynomial::Polynomial;
